refactor(whentime): rename cog package
